AntBrain names now need to start with a '~'
Fixed the lever. Was a problem distinguishing between comments / file
names and Instructions. Ant Brain names now start with a ‘~’  to deal
with this. Also, the lever and parser were looking for ‘PickUp’, the
string should have been ‘Pickup’.  Also, tweaked the regex and improved
robustness so the lever can deal with Food food FOOD etc. Plan to do
the same with the parser,